Fix a number of warnings - some cosmetic, some not

+++ KeyboardioScanner.h
@@ -70,6 +70,6 @@ class KeyboardioScanner {
     bool keyReady = false;
     byte nextLEDBank = 0;
     void sendLEDBank(byte bank);
-    int readRegister(int cmd);
+    int readRegister(uint8_t cmd);
 };
